Add script for preparing import of WordPress project into Local
The script addresses the issue that the location of the WordPress
project directory is hardcoded in the SQL database file. If then the
repository is cloned on a new machine to a different location and the
project is imported as is, this results in misbehaviour (in particular,
the URLs in the font CSS files are updated in a wrong way which results
in the font files not being found).

To solve this issue, this script performs a find-and-replace for the
hardcoded project directories in the SQL database file, replacing the
previous project directory with the new one.

This commit also already assumes the new location of the project
directory: namely the 'wordpress' directory in the repo instead of a
directory outside the repository. In this way, all changes are
automatically tracked by the repository and no separate syncing step is
necessary anymore before committing.